Adjust file splits, finish 186E8.c for psp
