+++ config/symbols.pspeu.dra.txt
@@ -18,6 +18,7 @@ SetNextRoomToLoad = 0x090E73D0;
 func_800F0CD8 = 0x090E74E8;
 func_800F298C = 0x090e9018;
 DebugShowWaitInfo = 0x090f5068;
+DebugInputWait = 0x090f50f8;
 DebugUpdate = 0x090f5180;
 MainGame = 0x090f5b58;
 HandleTitle = 0x090f6800;
