+++ src/dra_psp/18210.c
@@ -4,22 +4,3 @@
 INCLUDE_ASM("dra_psp/psp/dra_psp/18210", func_psp_090F4B90);
 
 INCLUDE_ASM("dra_psp/psp/dra_psp/18210", func_psp_090F4FC8);
-
-extern s32 g_DebugWaitInfoTimer;
-
-void DebugShowWaitInfo(const char* msg) {
-    g_CurrentBuffer = g_CurrentBuffer->next;
-    FntPrint(msg);
-    if (g_DebugWaitInfoTimer++ & 4) {
-        FntPrint("\no\n");
-    }
-    DrawSync(0);
-    VSync(0);
-    PutDrawEnv(&g_CurrentBuffer->draw);
-    PutDispEnv(&g_CurrentBuffer->disp);
-#if !defined(VERSION_PSP)
-    FntFlush(-1);
-#endif
-}
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/18210", func_psp_090F50F8);
